+++ src/layout/PrivateLayout.tsx
@@ -1,12 +1,12 @@
-import { useAuth } from "@src/auth/provider/AuthContext"
+import { useAuth } from "@src/context/AuthContext"
 import Login from "@src/components/ui/Login"
-import {  Outlet } from "react-router"
+import { Outlet } from "react-router"
 
 const PrivateLayout = () => {
 
 	const { session } = useAuth()
 
-	return session ? <Outlet/> : <Login/>
+	return session ? <Outlet /> : <Login />
 }
 
 export default PrivateLayout
