--- src/layout/index.ts
@@ -1,2 +1,0 @@
-export { default as Layout } from './Layout';
-export {default as PrivateLayout} from "@src/layout/PrivateLayout"
